+++ src/index.ts
@@ -3,8 +3,10 @@
  */
 export { EdvClient } from './EdvClient.js'
 export { EdvDocument } from './EdvDocument.js'
-export { EdvClientCore } from './EdvClientCore.js'
-export { EdvDocumentCipher } from './EdvDocumentCipher.js'
-export { assertDocId } from './assert.js'
 export { HttpsTransport } from './HttpsTransport.js'
-export { Transport } from './Transport.js'
+export {
+  EdvClientCore,
+  EdvDocumentCipher,
+  assertDocId,
+  Transport
+} from './core.js'